Set FileScanResult to Clean in localtest.
A better approach would be to wait 1 minute before setting it to clean, and provide some sort of known test file that always is `Infected`, but this simple fix makes it possible to use EnableFileScan locally.

The current code comes from https://github.com/Altinn/app-localtest/pull/46 which does not seem to deliberatly force localtest users to not use filescan.